+++ plugins/catalog-backend-module-unprocessed/src/index.ts
@@ -21,4 +21,4 @@
  */
 
 export * from './module';
-export * from './plugin';
+export * from './UnprocessedEntitiesModule';
